Fix smart contract compilation issues
- Change u8 to u32 for evidence_required (u8 not supported in Soroban)
- Replace floating point math with integer math for GPS distance calculation
- Fix symbol_short! macros to use string literals instead of constants
- Remove conflicting token module from reward-contract
- All 3 core contracts now compile successfully to WASM

Compiled contracts:
- certificate-nft.wasm (12.7KB)
- mission-contract (warnings only, compiles)
- reward-contract (compiles successfully)

+++ contracts/mission-contract/src/types.rs
@@ -31,7 +31,7 @@ pub struct Mission {
     pub status: MissionStatus,
     pub max_claimers: u32,        // Maximum number of people who can claim
     pub current_claimers: u32,    // Current number of claims
-    pub evidence_required: u8,    // Number of photos required
+    pub evidence_required: u32,   // Number of photos required
     pub deadline: u64,            // Unix timestamp
     pub metadata_uri: String,     // IPFS URI with description, images, etc.
     pub funded_amount: i128,      // Total amount funded
